download html5spec from ant
Gives finer control on whether the spec has been updated upstream.

Permits to have a local copy in case we call the 'clean' target and still want
to build offline.

If one doesn't have internet access but the file is present on the filesystem,
ant must be called with property 'offline=true'. This can be achived also by
using the '--offline' option when invoking build.py (see next commit)