Only run language detector on documents with sufficient text content
This change makes the language detector avoid checking any documents
that don’t have a “sufficient” amount of text content to base the
language-detection on. Currently, this change causes it to skip any
documents from which less than 1024 characters (about 150–200 words) of
“checkable” text can be extracted.

Otherwise, without this change, when the language detector is run on a
document that has very little text content, it can often unexpectedly
make surprisingly-wrong guesses about the language of a document.