ant: Build a bundle in the form that Maven Central expects/requires
- use “zip” task, not “jar” — because Maven doesn’t want a META-INF dir
- use/zip-up nu/validator/validator/${version.number} directory structure
- use the “checksum" task (not python) to create the md5 and sha1 checksums rather than in python)

Also note: this makes the “validator-bundle” target no longer depend on the
“validator-artifacts” target — because we want to run things in this order:

1. Run “ant validator-artifacts”	(to build the artifacts)
2. Run “python maven-sign”		(to gpg-sign the artifacts)
3. Run “ant validator-bundle”		(to build the Maven *-bundle.jar)

Step 2 (gpg signing) is something we can’t do with ant itself (because
ant doesn’t provide a good way to do it that we can use). Instead, we
have to do it from Python (or somewhere else that calls “gpg --sign”
from a subprocess). So if we were to make “validator-bundle“ depend on
“validator-artifacts”, then ant would (re)run the “validator-artifacts”
step — and thus re-generate all the files we had just already signed.

Also, when running the validator-bundle target, make the build/dist/nu
directory if it doesn’t already exist. That enables the validator-bundle
target to be run/tested on its own — even if the validator-artifacts
target has not been run. (Otherwise, if the validator-artifacts target
*has* been run, then the build/dist/nu structure will already exist.)