tools: Enable calling vnu from checker.py, for file/URL checking
This change provides you with the convenience of being able to check
documents just by doing this:

   ./checker.py check file.html https://example.com ...

That is, you invoke checker.py with the “check” argument — and after
that, provide any number of filenames or URLs of documents to check. The
checker.py script then runs vnu on all the files and URLs you specified.