ci(maven): Only run the maven-release job if we actually have a new version
This change causes the maven-release steps to be skipped if we’ve
already published a maven package with the same version number as what
we would be publishing if we did actually run all the steps.

The version-checking logic is in the build/build.py script, and CI uses
that by running “checker.py maven-version-exists” (newly-added task), and
checking the exit code (0 if the package already exists, 1 if it doesn’t).