fix: Fix multi-valued-CSS-functions checking for “sizes” values
This change fixes how we deal with commas in “sizes” values we check.
The existing “split on commas” logic we had in the code basically didn’t
anticipate the case of multi-valued functions with commas in them.

But with this change, it does. We no longer split on such commas. And so
now, we no longer unexpectedly report spurious errors that we had been.

Fixes https://github.com/validator/validator/issues/1585